Use a better secret strategy for moves
This change ensures that each game is random and non-determinstic.
It introduces hashing the details of the game using the parameters like
a random number generated by the system, the player's address, the
coin side and the timestamp.

More importantly, we now calculate the winner by using the unicode value
of the given secret. In js it would `charCodeAt/1`.

In the future, we will support apps embedding app data in the secret to
improve the secret's entropy even more.