Use new custom error syntax
Using custom errors will help reduce the gas fees on happy paths and
non-happy paths. The downside here is that we have to express the guard
logic in the negative condition which could be anti-intuitive when reading
the contract. Hopefully, when `require` can revert custom errors, we can
update these changes to use that syntax instead.

Equally, this change updates the Wallet logic. It replaces `creditWallet`
API with `transfer`. This enforces the rule that you have to deposit
first before transferring or crediting another address. This is better
for the Coinflip app because we infer the serviceCharge as the
wallet balance after transferring to other players.