Pay service charge eagerly
Contrary to the initial thought, we will be paying the
service charge eagerly right after a game is being
concluded. This will remove the need for an extra
state variable to keep track of `totalWagers` or
similar state variable for defering the service
charge payment.

There major drawback of that deferring approach is that
we will be exposing all game wagers vs this approach
that only exposes a single game wager.