Abort coinflip games request when deps change
This is in a bid to handle Next js's useRouter API flaw. The
useRouter API initially returns an empty query object even if it exists
from the start. We handle this edge case by simply canceling the
previous request when Next js's useRouter returns the actual query
params. Since this happens in a split second, we reduce the extra
network overhead in the user's client.