Rename nonGameBalance to balance
nonGameBalance was initially introduced to curb the ambiguity that
stems from a player having a balance in a given game and a wallet balance.
That wallet strategy is now stale because, wagers mappings is the responsibility
of the game apps. The wallets simply helps authorize the games a given app
can control. With this approach in mind, a game balance is never ever attributed
to a player or wallet owner, but always to an app's game. This means we no longer
have to deal with the ambiguity of gameBalances vs nonGameBalances. Hence, the
reason we no longer need the descriptive prefix and simply call them 'balances'